refactor: move db functions that run queries to a separate file in the db package
